#135 create tests for event register

See merge request itatt1005_2024_group15/purchase-planner!87

+++ src/main/java/no/ntnu/idatt1002/demo/repo/EventRegister.java
@@ -4,10 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-
-import no.ntnu.idatt1002.demo.Logger;
 import no.ntnu.idatt1002.demo.dao.DAO;
 import no.ntnu.idatt1002.demo.data.Event;
+import no.ntnu.idatt1002.demo.util.VerifyInput;
 
 
 /**
@@ -22,28 +21,12 @@ public class EventRegister {
    * Constructor for the EventRegister class.
    *
    * @param dao the data access object
+   * @throws IllegalArgumentException if dao is null
    */
   public EventRegister(DAO dao) {
+    VerifyInput.verifyNotNull(dao, "dao");
     this.dao = dao;
     this.events = new ArrayList<>();
-  }
-
-  /**
-   * This method returns the events in the register.
-   *
-   * @return the events in the register as a Map
-   */
-  public Map<Integer, Event> getEvents() {
-    return events.stream().collect(Collectors.toMap(Event::getId, event -> event));
-  }
-
-  /**
-   * This method returns the events in the register as a list.
-   *
-   * @return the events in the register as a list
-   */
-  public List<Event> getEventsAsList() {
-    return events;
   }
 
   /**
@@ -51,7 +34,7 @@ public class EventRegister {
    */
   public void getAllEvents() {
     events = new ArrayList<>();
-    List<List<String>> events = dao.getAllFromTable("Recipe", "Event", "recipe_id");
+    List<List<String>> events = dao.getAllFromTable("Event", "Recipe", "recipe_id");
     packageToEvent(events);
   }
 
@@ -72,7 +55,7 @@ public class EventRegister {
    * @param id the id of the event
    * @return the index of the event
    */
-  private int getIdFromEvent(int id) {
+  private int getIndexFromEvent(int id) {
     for (int i = 0; i < events.size(); i++) {
       if (events.get(i).getId() == id) {
         return i;
@@ -85,10 +68,35 @@ public class EventRegister {
    * This method deletes an event from the database.
    *
    * @param id the id of the event to be deleted
+   * @throws IllegalArgumentException if the event does not exist
    */
   public void deleteEvent(int id) {
-    int index = getIdFromEvent(id);
+    int index = getIndexFromEvent(id);
+    if (index == -1) {
+      throw new IllegalArgumentException("Event with id " + id + " does not exist.");
+    }
     dao.deleteFromDatabase(events.get(index));
+  }
+
+  /**
+   * This method gets the events by the recipe id.
+   *
+   * @param id the id of the recipe
+   */
+  public void getEventById(int id) {
+    events = new ArrayList<>();
+    List<List<String>> events = dao.filterFromTable(
+        "Event", "event_id", Integer.toString(id), "Recipe", "recipe_id");
+    packageToEvent(events);
+  }
+
+  /**
+   * This method returns the events in the register as a list.
+   *
+   * @return the events in the register as a list
+   */
+  public List<Event> getEventsAsList() {
+    return events;
   }
 
   /**
@@ -97,8 +105,12 @@ public class EventRegister {
    * @param event_id the id of the event
    * @param recipe_id the id of the recipe
    * @param date the date of the event
+   * @throws IllegalArgumentException if the event does not exist
    */
   public void updateEvent(int event_id, int recipe_id, int date) {
+    if (getIndexFromEvent(event_id) == -1) {
+      throw new IllegalArgumentException("Event with id " + event_id + " does not exist.");
+    }
     Event event = new Event(event_id, recipe_id, date);
     dao.updateDatabase(event);
   }
@@ -109,19 +121,10 @@ public class EventRegister {
    * @param date the date of the event
    */
   public void getEventsByDate(int date) {
+    VerifyInput.verifyDateLength(date, "date");
     events = new ArrayList<>();
-    List<List<String>> events = dao.filterFromTable("Recipe", "date", Integer.toString(date), "Event", "recipe_id");
-    packageToEvent(events);
-  }
-
-  /**
-   * This method gets the events by the recipe id.
-   *
-   * @param id the id of the recipe
-   */
-  public void getEventById(int id) {
-    events = new ArrayList<>();
-    List<List<String>> events = dao.filterFromTable("Recipe", "event_id", Integer.toString(id), "Event", "recipe_id");
+    List<List<String>> events = dao.filterFromTable(
+        "Event", "date", Integer.toString(date), "recipe", "recipe_id");
     packageToEvent(events);
   }
 
@@ -133,17 +136,25 @@ public class EventRegister {
   private void packageToEvent(List<List<String>> eventList) {
     List<Event> newEvents = new ArrayList<>();
     for (List<String> event : eventList) {
-      Logger.warning(event.toString());
       newEvents.add(new Event(
-          Integer.parseInt(event.get(4)),
-          Integer.parseInt(event.get(5)),
-          event.get(1),
-          Integer.parseInt(event.get(6)),
-          event.get(3),
-          Integer.parseInt(event.get(2))
+          Integer.parseInt(event.get(0)),
+          Integer.parseInt(event.get(1)),
+          event.get(4),
+          Integer.parseInt(event.get(2)),
+          event.get(6),
+          Integer.parseInt(event.get(5))
       ));
     }
     this.events = newEvents;
+  }
+
+  /**
+   * This method returns the events in the register.
+   *
+   * @return the events in the register as a Map
+   */
+  public Map<Integer, Event> getEvents() {
+    return events.stream().collect(Collectors.toMap(Event::getId, event -> event));
   }
 
   /**
